Crate subxt

Subxt is a library to submit extrinsics to a substrate node via RPC.

The generated Subxt API exposes the ability to:

Submit extrinsics (Calls)

Query storage (Storage)

Query constants (Constants)

Subscribe to events (Events)

Generate the runtime API

Subxt generates a runtime API from downloaded static metadata. The metadata can be downloaded using the subxt-cli tool.

To generate the runtime API, use the subxt attribute which points at downloaded static metadata.

#[subxt::subxt(runtime_metadata_path = "metadata.scale")]
pub mod node_runtime { }

The node_runtime has the following hierarchy:

pub mod node_runtime {
    pub mod PalletName {
        pub mod calls { }
        pub mod storage { }
        pub mod constants { }
        pub mod events { }
    }
}

For more information regarding the node_runtime hierarchy, please visit the subxt-codegen documentation.

Initializing the API client

use subxt::{ClientBuilder, DefaultConfig, PolkadotExtrinsicParams};

#[subxt::subxt(runtime_metadata_path = "../artifacts/polkadot_metadata.scale")]
pub mod polkadot {}

let api = ClientBuilder::new()
    .set_url("wss://rpc.polkadot.io:443")
    .build()
    .await
    .unwrap()
    .to_runtime_api::<polkadot::RuntimeApi<DefaultConfig, PolkadotExtrinsicParams<DefaultConfig>>>();

The RuntimeApi type is generated by the subxt macro from the supplied metadata. This can be parameterized with user supplied implementations for the Config and Extra types, if the default implementation differs from the target chain.

To ensure that extrinsics are properly submitted, during the build phase of the Client the runtime metadata of the node is downloaded. If the URL is not specified (set_url), the local host is used instead.

Submit Extrinsics

Extrinsics are obtained using the API’s RuntimeApi::tx() method, followed by pallet_name() and then the call_item_name().

Submit an extrinsic, returning success once the transaction is validated and accepted into the pool:

Please visit the balance_transfer example for more details.

Querying Storage

The runtime storage is queried via the generated RuntimeApi::storage() method, followed by the pallet_name() and then the storage_item_name().

Please visit the fetch_staking_details example for more details.

Query Constants

Constants are embedded into the node’s metadata.

The subxt offers the ability to query constants from the runtime metadata (metadata downloaded when constructing the client, not the one provided for API generation).

To query constants use the generated RuntimeApi::constants() method, followed by the pallet_name() and then the constant_item_name().

Please visit the fetch_constants example for more details.

Subscribe to Events

To subscribe to events, use the generated RuntimeApi::events() method which exposes:

subscribe() - Subscribe to events emitted from blocks. These blocks haven’t necessarily been finalised.

subscribe_finalized() - Subscribe to events from finalized blocks.

at() - Obtain events at a given block hash.

Examples

subscribe_all_events: Subscribe to events emitted from blocks.

subscribe_one_event: Subscribe and filter by one event.

subscribe_some_events: Subscribe and filter event.

Static Metadata Validation

There are two types of metadata that the subxt is aware of:

static metadata: Metadata used for generating the API.

runtime metadata: Metadata downloaded from the target node when a subxt client is created.

There are cases when the static metadata is different from the runtime metadata of a node. Such is the case when the node performs a runtime update.

To ensure that subxt can properly communicate with the target node the static metadata is validated against the runtime metadata of the node.

This validation is performed at the Call, Constant, and Storage levels, as well for the entire metadata. The level of granularity ensures that the users can still submit a given call, even if another call suffered changes.

Full metadata validation:

// To make sure that all of our statically generated pallets are compatible with the
// runtime node, we can run this check:
api.validate_metadata().unwrap();

Call level validation:

// Submit the `transfer` extrinsic from Alice's account to Bob's.
let dest = AccountKeyring::Bob.to_account_id().into();

let extrinsic = api
    .tx()
    .balances()
    // Constructing an extrinsic will fail if the metadata
    // is not in sync with the generated API.
    .transfer(dest, 123_456_789_012_345)
    .unwrap();

Runtime Updates

There are cases when the node would perform a runtime update, and the runtime node’s metadata would be out of sync with the subxt’s metadata.

The UpdateClient API keeps the RuntimeVersion and Metadata of the client synced with the target node.

Please visit the subscribe_runtime_updates example for more details.

Re-exports

pub use bitvec;

pub use codec;

pub use sp_core;

pub use sp_runtime;

pub use crate::rpc::BlockNumber;

pub use crate::rpc::ReadProof;

pub use crate::rpc::SystemProperties;

pub use crate::storage::KeyIter;

pub use crate::storage::StorageEntry;

pub use crate::storage::StorageEntryKey;

pub use crate::storage::StorageMapKey;

Modules

events

This module exposes the ability to work with events generated by a given block. Subxt can either attempt to statically decode events into known types, or it can hand back details of the raw event without knowing what shape its contents are (this may be useful if we don’t know what exactly we’re looking for).

extrinsic

Create signed or unsigned extrinsics.

rpc

RPC types and client for interacting with a substrate node.

storage

Query the runtime storage using StorageClient.

updates

Perform runtime updates in the background using UpdateClient.

Structs

CallData

Decoded call data and associated type information.

Client

Client to interface with a substrate node.

ClientBuilder

ClientBuilder for constructing a Client.

Decoder

Decoder

DecoderBuilder

Pallet name

Encoded

Wraps an already encoded byte vector, prevents being encoded as a raw byte vector as part of the transaction payload

ErrorMetadata

Metadata for specific errors obtained from the pallet’s PalletErrorMetadata.

EventDetails

A decoded event and associated details.

Events

A collection of events obtained from a block, bundled with the necessary information needed to decode and iterate over them.

Extrinsic

The result of successfully decoding an extrinsic.

Metadata

A representation of the runtime metadata received from a node.

ModuleError

Details about a module error that has occurred.

ModuleErrorData

The error details about a module error that has occurred.

PairSigner

A Signer implementation that can be constructed from an Pair.

PalletMetadata

Metadata for a specific pallet.

PathKey

Path key

RawEventDetails

The raw bytes for an event with associated details about where and when it was emitted.

RpcClient

Generic asynchronous client.

RuntimeError

This is used in the place of the E in GenericError<E> when we may have a Runtime Error. We use this wrapper so that it is possible to implement From<Error<Infallible> for Error<RuntimeError<E>>.

SubmittableExtrinsic

A constructed call ready to be signed and submitted.

TransactionEvents

This represents the events related to our transaction. We can iterate over the events, or look for a specific one.

TransactionInBlock

This struct represents a transaction that has made it into a block.

TransactionProgress

This struct represents a subscription to the progress of some transaction, and is returned from crate::SubmittableExtrinsic::sign_and_submit_then_watch().

WrapperKeepOpaque

A wrapper for any type T which implement encode/decode in a way compatible with Vec<u8>.

Enums

DefaultConfig

Default set of commonly used types by Substrate runtimes.

GenericError

The underlying error enum, generic over the type held by the Runtime variant. Prefer to use the Error<E> and BasicError aliases over using this type directly.

MetadataError

Metadata error originated from inspecting the internal representation of the runtime metadata.

Phase

A phase of a block’s execution.

StorageHasher

Hasher used by storage maps

TransactionError

Transaction error.

TransactionStatus

Possible transaction statuses returned from our TransactionProgress::next_item() call.

Traits

Call

Trait to uniquely identify the call (extrinsic)’s identity from the runtime metadata.

Config

Runtime types.

Event

Trait to uniquely identify the events’s identity from the runtime metadata.

HasModuleError

This trait is automatically implemented for the generated DispatchError, so that we can pluck out information about the Module error variant, if` it exists.

Type Definitions

BasicError

An error that will never contain a runtime error.

Error

An error that may contain some runtime error E

KeyedVec

This represents a key-value collection and is SCALE compatible with collections like BTreeMap. This has the same type params as BTreeMap which allows us to easily swap the two during codegen.

PolkadotExtrinsicParams

A struct representing the signed extra and additional parameters required to construct a transaction for a polkadot node.

PolkadotExtrinsicParamsBuilder

A builder which leads to PolkadotExtrinsicParams being constructed. This is what you provide to methods like sign_and_submit().

SubstrateExtrinsicParams

A struct representing the signed extra and additional parameters required to construct a transaction for the default substrate node.

SubstrateExtrinsicParamsBuilder

A builder which leads to SubstrateExtrinsicParams being constructed. This is what you provide to methods like sign_and_submit().

Attribute Macros

subxt