Module subxt::config

This module provides a Config type, which is used to define various types that are important in order to speak to a particular chain. SubstrateConfig provides a default set of these types suitable for the default Substrate node implementation, and PolkadotConfig for a Polkadot node.

Structs

WithExtrinsicParams

Take a type implementing Config (eg SubstrateConfig), and some type which describes the additional and extra parameters to pass to an extrinsic (see crate::tx::ExtrinsicParams), and returns a type implementing Config with those new ExtrinsicParams.

Enums

SubstrateConfig

Default set of commonly used types by Substrate runtimes.

Traits

Config

Runtime types.

Parameter

Parameter trait copied from substrate::frame_support

Type Definitions

PolkadotConfig

Default set of commonly used types by Polkadot nodes.