Module subxt::config::substrate

Substrate specific configuration

Re-exports

pub use crate::utils::account_id::AccountId32;

pub use crate::utils::multi_address::MultiAddress;

pub use crate::utils::multi_signature::MultiSignature;

pub use super::extrinsic_params::Era;

Structs

AssetTip

A tip payment made in the form of a specific asset.

BlakeTwo256

A type that can hash values using the blaks2_256 algorithm.

Digest

Generic header digest. From sp_runtime::generic::digest.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

SubstrateHeader

A generic Substrate header type, adapted from sp_runtime::generic::Header. The block number and hasher can be configured to adapt this for other nodes.

U256

Little-endian large integer type 256-bit unsigned integer.

Enums

DigestItem

Digest item that is able to encode/decode ‘system’ digest items and provide opaque access to other items. From sp_runtime::generic::digest.

SubstrateConfig

Default set of commonly used types by Substrate runtimes.

Type Definitions

ConsensusEngineId

Consensus engine unique ID. From sp_runtime::ConsensusEngineId.

SubstrateExtrinsicParams

A struct representing the signed extra and additional parameters required to construct a transaction for the default substrate node.

SubstrateExtrinsicParamsBuilder

A builder which leads to SubstrateExtrinsicParams being constructed. This is what you provide to methods like sign_and_submit().