Trait subxt::config::Config

pub trait Config: 'static {
    type Index: Parameter + Member + DeserializeOwned + Default + AtLeast32Bit + Copy + TypeInfo + Into<u64>;
    type BlockNumber: Parameter + Member + Default + Copy + Hash + FromStr + Into<u64>;
    type Hash: Parameter + Member + MaybeSerializeDeserialize + Ord + Default + Copy + Hash + AsRef<[u8]> + AsMut<[u8]> + TypeInfo;
    type Hashing: Hash<Output = Self::Hash>;
    type AccountId: Parameter + Member + Serialize;
    type Address: Codec + Clone + PartialEq;
    type Header: Parameter + Header<Number = Self::BlockNumber, Hash = Self::Hash> + DeserializeOwned;
    type Signature: Verify + Encode + Send + Sync + 'static;
    type Extrinsic: Parameter + Extrinsic + Debug + MaybeSerializeDeserialize;
    type ExtrinsicParams: ExtrinsicParams<Self::Index, Self::Hash>;
}

Runtime types.

Required Associated Types

type Index: Parameter + Member + DeserializeOwned + Default + AtLeast32Bit + Copy + TypeInfo + Into<u64>

Account index (aka nonce) type. This stores the number of previous transactions associated with a sender account.

type BlockNumber: Parameter + Member + Default + Copy + Hash + FromStr + Into<u64>

The block number type used by the runtime.

type Hash: Parameter + Member + MaybeSerializeDeserialize + Ord + Default + Copy + Hash + AsRef<[u8]> + AsMut<[u8]> + TypeInfo

The output of the Hashing function.

type Hashing: Hash<Output = Self::Hash>

The hashing system (algorithm) being used in the runtime (e.g. Blake2).

type AccountId: Parameter + Member + Serialize

The user account identifier type for the runtime.

type Address: Codec + Clone + PartialEq

The address type. This instead of <frame_system::Trait::Lookup as StaticLookup>::Source.

type Header: Parameter + Header<Number = Self::BlockNumber, Hash = Self::Hash> + DeserializeOwned

The block header.

type Signature: Verify + Encode + Send + Sync + 'static

Signature type.

type Extrinsic: Parameter + Extrinsic + Debug + MaybeSerializeDeserialize

Extrinsic type within blocks.

type ExtrinsicParams: ExtrinsicParams<Self::Index, Self::Hash>

This type defines the extrinsic extra and additional parameters.