Trait subxt::constants::ConstantAddress

pub trait ConstantAddress {
    type Target: DecodeWithMetadata;
    fn pallet_name(&self) -> &str;
    fn constant_name(&self) -> &str;

    fn validation_hash(&self) -> Option<[u8; 32]> { ... }
}

This represents a constant address. Anything implementing this trait can be used to fetch constants.

Associated Types

type Target: DecodeWithMetadata

The target type of the value that lives at this address.

Required methods

fn pallet_name(&self) -> &str

The name of the pallet that the constant lives under.

fn constant_name(&self) -> &str

The name of the constant in a given pallet.

Provided methods

fn validation_hash(&self) -> Option<[u8; 32]>

An optional hash which, if present, will be checked against the node metadata to confirm that the return type matches what we are expecting.