Module subxt::events

This module exposes the ability to work with events generated by a given block. Subxt can either attempt to statically decode events into known types, or it can hand back details of the raw event without knowing what shape its contents are (this may be useful if we don’t know what exactly we’re looking for).

This module is wrapped by the generated API in RuntimeAPI::EventsApi.

Examples

Subscribe to all events

Users can subscribe to all emitted events from blocks using subscribe().

To subscribe to all events from just the finalized blocks use subscribe_finalized().

To obtain the events from a given block use at().

let mut events = api.events().subscribe().await.unwrap();

while let Some(ev) = events.next().await {
    // Obtain all events from this block.
    let ev: subxt::Events<_, _> = ev.unwrap();
    // Print block hash.
    println!("Event at block hash {:?}", ev.block_hash());
    // Iterate over all events.
    let mut iter = ev.iter();
    while let Some(event_details) = iter.next() {
        println!("Event details {:?}", event_details);
    }
}

Filter events

The subxt exposes the ability to filter events via the filter_events() function.

The function filters events from the provided tuple. If 1-tuple is provided, the events are returned directly. Otherwise, we’ll be given a corresponding tuple of Option’s, with exactly one variant populated each time.

#[subxt::subxt(runtime_metadata_path = "../artifacts/polkadot_metadata.scale")]
pub mod polkadot {}

let mut transfer_events = api
    .events()
    .subscribe()
    .await
    .unwrap()
    .filter_events::<(polkadot::balances::events::Transfer,)>();

while let Some(transfer_event) = transfer_events.next().await {
    println!("Balance transfer event: {transfer_event:?}");
}

Structs

EventDetails

A decoded event and associated details.

EventSubscription

A subscription to events that implements Stream, and returns Events objects for each block.

Events

A collection of events obtained from a block, bundled with the necessary information needed to decode and iterate over them.

FilterEvents

A stream which filters events based on the Filter param provided. If Filter is a 1-tuple of a single Event type, it will return every instance of that event as it’s found. If filter is tuple of multipleEventtypes, it will return a corresponding tuple ofOptions, where exactly one of these will be Some(event)` each iteration.

FilteredEventDetails

This is returned from the FilterEvents impl of Stream. It contains some type representing an event we’ve filtered on, along with couple of additional pieces of information about that event.

RawEventDetails

The raw bytes for an event with associated details about where and when it was emitted.

Traits

EventFilter

This trait is implemented for tuples of Event types; any such tuple (up to size 8) can be used to filter an event subscription to return only the specified events.

Type Definitions

DecodedValue

A Value which has been decoded from some raw bytes.