Struct subxt::events::FilterEvents

pub struct FilterEvents<'a, Sub: 'a, T: Config, Filter: EventFilter> { /* private fields */ }

A stream which filters events based on the Filter param provided. If Filter is a 1-tuple of a single Event type, it will return every instance of that event as it’s found. If filter is tuple of multipleEventtypes, it will return a corresponding tuple ofOptions, where exactly one of these will be Some(event)` each iteration.