Trait subxt::events::EventFilter

pub trait EventFilter: Sealed {
    type ReturnType;

    fn filter<T: Config>(
        events: Events<T>
    ) -> Box<dyn Iterator<Item = Result<FilteredEventDetails<T::Hash, Self::ReturnType>, Error>> + Send>;
}

This trait is implemented for tuples of Event types; any such tuple (up to size 8) can be used to filter an event subscription to return only the specified events.

Required Associated Types

type ReturnType

The type we’ll be handed back from filtering.

Required Methods

fn filter<T: Config>(
events: Events<T>
) -> Box<dyn Iterator<Item = Result<FilteredEventDetails<T::Hash, Self::ReturnType>, Error>> + Send>

Filter the events based on the type implementing this trait.