Module subxt::extrinsic

Create signed or unsigned extrinsics.

Structs

AssetTip

A tip payment made in the form of a specific asset.

PairSigner

A Signer implementation that can be constructed from an Pair.

PlainTip

A tip payment.

Enums

Era

An era to describe the longevity of a transaction.

Traits

ExtrinsicParams

This trait allows you to configure the “signed extra” and “additional” parameters that are signed and used in transactions. see [BaseExtrinsicParams] for an implementation that is compatible with a Polkadot node.

Signer

Signing transactions requires a Signer. This is responsible for providing the “from” account that the transaction is being signed by, as well as actually signing a SCALE encoded payload. Optionally, a signer can also provide the nonce for the transaction to use.

Type Definitions

PolkadotExtrinsicParams

A struct representing the signed extra and additional parameters required to construct a transaction for a polkadot node.

PolkadotExtrinsicParamsBuilder

A builder which leads to PolkadotExtrinsicParams being constructed. This is what you provide to methods like sign_and_submit().

SubstrateExtrinsicParams

A struct representing the signed extra and additional parameters required to construct a transaction for the default substrate node.

SubstrateExtrinsicParamsBuilder

A builder which leads to SubstrateExtrinsicParams being constructed. This is what you provide to methods like sign_and_submit().