Struct subxt::rpc::Rpc

pub struct Rpc<T: Config> { /* private fields */ }

Client for substrate rpc interfaces

Implementations

impl<T: Config> Rpc<T>

pub fn new<R: RpcClientT>(client: Arc<R>) -> Self

Create a new Rpc

pub async fn storage(
&self,
key: &[u8],
hash: Option<T::Hash>
) -> Result<Option<StorageData>, Error>

Fetch the raw bytes for a given storage key

pub async fn storage_keys_paged(
&self,
key: &[u8],
count: u32,
start_key: Option<&[u8]>,
hash: Option<T::Hash>
) -> Result<Vec<StorageKey>, Error>

Returns the keys with prefix with pagination support. Up to count keys will be returned. If start_key is passed, return next keys in storage in lexicographic order.

pub async fn query_storage(
&self,
keys: impl IntoIterator<Item = &[u8]>,
from: T::Hash,
to: Option<T::Hash>
) -> Result<Vec<StorageChangeSet<T::Hash>>, Error>

Query historical storage entries

pub async fn query_storage_at(
&self,
keys: impl IntoIterator<Item = &[u8]>,
at: Option<T::Hash>
) -> Result<Vec<StorageChangeSet<T::Hash>>, Error>

Query historical storage entries

pub async fn genesis_hash(&self) -> Result<T::Hash, Error>

Fetch the genesis hash

pub async fn metadata(&self, at: Option<T::Hash>) -> Result<Metadata, Error>

Fetch the metadata

pub async fn call(
&self,
function: String,
call_parameters: Option<&[u8]>,
at: Option<T::Hash>
) -> Result<Bytes, Error>

Execute a runtime API call.

pub async fn system_properties(&self) -> Result<SystemProperties, Error>

Fetch system properties

pub async fn system_health(&self) -> Result<Health, Error>

Fetch system health

pub async fn system_chain(&self) -> Result<String, Error>

Fetch system chain

pub async fn system_name(&self) -> Result<String, Error>

Fetch system name

pub async fn system_version(&self) -> Result<String, Error>

Fetch system version

pub async fn system_account_next_index<AccountId: Serialize>(
&self,
account: &AccountId
) -> Result<T::Index, Error>

Fetch the current nonce for the given account ID.

pub async fn header(
&self,
hash: Option<T::Hash>
) -> Result<Option<T::Header>, Error>

Get a header

pub async fn block_hash(
&self,
block_number: Option<BlockNumber>
) -> Result<Option<T::Hash>, Error>

Get a block hash, returns hash of latest block by default

pub async fn finalized_head(&self) -> Result<T::Hash, Error>

Get a block hash of the latest finalized block

pub async fn block(
&self,
hash: Option<T::Hash>
) -> Result<Option<ChainBlockResponse<T>>, Error>

Get a Block

pub async fn block_stats(
&self,
block_hash: T::Hash
) -> Result<Option<BlockStats>, Error>

Reexecute the specified block_hash and gather statistics while doing so.

This function requires the specified block and its parent to be available at the queried node. If either the specified block or the parent is pruned, this function will return None.

pub async fn read_proof(
&self,
keys: impl IntoIterator<Item = &[u8]>,
hash: Option<T::Hash>
) -> Result<ReadProof<T::Hash>, Error>

Get proof of storage entries at a specific block’s state.

pub async fn runtime_version(
&self,
at: Option<T::Hash>
) -> Result<RuntimeVersion, Error>

Fetch the runtime version

pub async fn subscribe_best_block_headers(
&self
) -> Result<Subscription<T::Header>, Error>

Subscribe to all new best block headers.

pub async fn subscribe_all_block_headers(
&self
) -> Result<Subscription<T::Header>, Error>

Subscribe to all new block headers.

pub async fn subscribe_finalized_block_headers(
&self
) -> Result<Subscription<T::Header>, Error>

Subscribe to finalized block headers.

Note: this may not produce every block in the finalized chain; sometimes multiple blocks are finalized at once, and in this case only the latest one is returned. the higher level APIs that use this “fill in” the gaps for us.

pub async fn subscribe_runtime_version(
&self
) -> Result<Subscription<RuntimeVersion>, Error>

Subscribe to runtime version updates that produce changes in the metadata.

pub async fn submit_extrinsic<X: Encode>(
&self,
extrinsic: X
) -> Result<T::Hash, Error>

Create and submit an extrinsic and return corresponding Hash if successful

pub async fn state_call(
&self,
function: &str,
call_parameters: Option<&[u8]>,
at: Option<T::Hash>
) -> Result<Bytes, Error>

Execute a runtime API call.

pub async fn watch_extrinsic<X: Encode>(
&self,
extrinsic: X
) -> Result<Subscription<SubstrateTxStatus<T::Hash, T::Hash>>, Error>

Create and submit an extrinsic and return a subscription to the events triggered.

pub async fn insert_key(
&self,
key_type: String,
suri: String,
public: Bytes
) -> Result<(), Error>

Insert a key into the keystore.

pub async fn rotate_keys(&self) -> Result<Bytes, Error>

Generate new session keys and returns the corresponding public keys.

pub async fn has_session_keys(&self, session_keys: Bytes) -> Result<bool, Error>

Checks if the keystore has private keys for the given session public keys.

session_keys is the SCALE encoded session keys object from the runtime.

Returns true iff all private keys could be found.

pub async fn has_key(
&self,
public_key: Bytes,
key_type: String
) -> Result<bool, Error>

Checks if the keystore has private keys for the given public key and key type.

Returns true if a private key could be found.

pub async fn dry_run(
&self,
encoded_signed: &[u8],
at: Option<T::Hash>
) -> Result<DryRunResult, Error>

Submits the extrinsic to the dry_run RPC, to test if it would succeed.

Returns a types::DryRunResult, which is the result of performing the dry run.

pub async fn chainhead_unstable_follow(
&self,
runtime_updates: bool
) -> Result<Subscription<FollowEvent<T::Hash>>, Error>

Subscribe to chainHead_unstable_follow to obtain all reported blocks by the chain.

The subscription ID can be used to make queries for the block’s body (chainhead_unstable_body), block’s header (chainhead_unstable_header), block’s storage (chainhead_unstable_storage) and submitting runtime API calls at this block (chainhead_unstable_call).

Note

When the user is no longer interested in a block, the user is responsible for calling the chainhead_unstable_unpin method. Failure to do so will result in the subscription being stopped by generating the Stop event.

pub async fn chainhead_unstable_body(
&self,
subscription_id: String,
hash: T::Hash
) -> Result<Subscription<ChainHeadEvent<String>>, Error>

Subscribe to chainHead_unstable_body to obtain events regarding the block’s body.

Note

The subscription ID is obtained from an open subscription created by chainhead_unstable_follow.

pub async fn chainhead_unstable_header(
&self,
subscription_id: String,
hash: T::Hash
) -> Result<Option<String>, Error>

Get the block’s body using the chainHead_unstable_header method.

Note

The subscription ID is obtained from an open subscription created by chainhead_unstable_follow.

pub async fn chainhead_unstable_storage(
&self,
subscription_id: String,
hash: T::Hash,
key: &[u8],
child_key: Option<&[u8]>
) -> Result<Subscription<ChainHeadEvent<Option<String>>>, Error>

Subscribe to chainHead_storage to obtain events regarding the block’s storage.

Note

The subscription ID is obtained from an open subscription created by chainhead_unstable_follow.

pub async fn chainhead_unstable_call(
&self,
subscription_id: String,
hash: T::Hash,
function: String,
call_parameters: &[u8]
) -> Result<Subscription<ChainHeadEvent<String>>, Error>

Subscribe to chainHead_call to obtain events regarding the runtime API call.

Note

The subscription ID is obtained from an open subscription created by chainhead_unstable_follow.

pub async fn chainhead_unstable_unpin(
&self,
subscription_id: String,
hash: T::Hash
) -> Result<(), Error>

Unpin a block reported by the chainHead_follow subscription.

Note

The subscription ID is obtained from an open subscription created by chainhead_unstable_follow.

pub async fn chainhead_unstable_genesishash(&self) -> Result<T::Hash, Error>

Get genesis hash obtained from the chainHead_genesisHash method.