Module subxt::storage

Query the runtime storage using StorageClient.

This module is the core of performing runtime storage queries. While you can work with it directly, it’s prefer to use the generated storage() interface where possible.

The exposed API is performing RPC calls to state_getStorage and state_getKeysPaged.

A runtime storage entry can be of type:

StorageEntryKey::Plain for keys constructed just from the prefix twox_128(pallet) ++ twox_128(storage_item)

StorageEntryKey::Map for mapped keys constructed from the prefix, plus other arguments twox_128(pallet) ++ twox_128(storage_item) ++ hash(arg1) ++ arg1

Examples

Fetch Storage Keys

#[subxt::subxt(runtime_metadata_path = "../artifacts/polkadot_metadata.scale")]
pub mod polkadot {}

// Fetch just the keys, returning up to 10 keys.
let keys = storage
    .fetch_keys::<polkadot::xcm_pallet::storage::VersionNotifiers>(10, None, None)
    .await
    .unwrap();
// Iterate over each key
for key in keys.iter() {
    println!("Key: 0x{}", hex::encode(&key));
}

Iterate over Storage

#[subxt::subxt(runtime_metadata_path = "../artifacts/polkadot_metadata.scale")]
pub mod polkadot {}

// Iterate over keys and values.
let mut iter = storage
    .iter::<polkadot::xcm_pallet::storage::VersionNotifiers>(None)
    .await
    .unwrap();
while let Some((key, value)) = iter.next().await.unwrap() {
    println!("Key: 0x{}", hex::encode(&key));
    println!("Value: {}", value);
}

Structs

KeyIter

Iterates over key value pairs in a map.

StorageClient

Client for querying runtime storage.

StorageKeyPrefix

The prefix of the key to a StorageEntry

StorageMapKey

Storage key for a Map.

Enums

StorageEntryKey

Storage key.

Traits

SignedExtension

Means by which a transaction may be extended. This type embodies both the data and the logic that should be additionally associated with the transaction. It should be plain old data.

StorageEntry

Storage entry trait.