Module subxt::storage::address

Types representing an address which describes where a storage entry lives and how to properly decode it.

Structs

DynamicStorageAddress

This represents a dynamically generated storage address.

StaticStorageAddress

This represents a statically generated storage lookup address.

StorageMapKey

Storage key for a Map.

Yes

Used to signal whether a StorageAddress can be iterated, fetched and returned with a default value in the type system.

Enums

StorageHasher

Hasher used by storage maps

Traits

StorageAddress

This represents a storage address. Anything implementing this trait can be used to fetch and iterate over storage entries.

Functions

dynamic

Construct a new dynamic storage lookup.

dynamic_root

Construct a new dynamic storage lookup to the root of some entry.