Trait subxt::AccountData

pub trait AccountData {
    type StorageEntry: StorageEntry;
    type AccountId;
    type Index;
    fn storage_entry(account_id: Self::AccountId) -> Self::StorageEntry;

    fn nonce(
        result: &<Self::StorageEntry as StorageEntry>::Value
    ) -> Self::Index;
}

Trait to fetch data about an account.

Associated Types

type StorageEntry: StorageEntry

The runtime storage entry from which the account data can be fetched. Usually generated by the subxt macro.

type AccountId

The type of the account id to fetch the account data for.

type Index

The type of the account nonce returned from storage.

Required methods

fn storage_entry(account_id: Self::AccountId) -> Self::StorageEntry

Create a new storage entry key from the account id.

fn nonce(result: &<Self::StorageEntry as StorageEntry>::Value) -> Self::Index

Get the nonce from the storage entry value.