Trait subxt::tx::Signer

pub trait Signer<T: Config> {
    fn nonce(&self) -> Option<T::Index>;
    fn account_id(&self) -> &T::AccountId;
    fn address(&self) -> T::Address;
    fn sign(&self, signer_payload: &[u8]) -> T::Signature;
}

Signing transactions requires a Signer. This is responsible for providing the “from” account that the transaction is being signed by, as well as actually signing a SCALE encoded payload. Optionally, a signer can also provide the nonce for the transaction to use.

Required methods

fn nonce(&self) -> Option<T::Index>

Optionally returns a nonce.

fn account_id(&self) -> &T::AccountId

Return the “from” account ID.

fn address(&self) -> T::Address

Return the “from” address.

fn sign(&self, signer_payload: &[u8]) -> T::Signature

Takes a signer payload for an extrinsic, and returns a signature based on it.

Some signers may fail, for instance because the hardware on which the keys are located has refused the operation.