Module subxt::dynamic

This module provides the entry points to create dynamic transactions, storage and constant lookups.

Structs

Value

Value holds a representation of some value that has been decoded, as well as some arbitrary context.

Functions

constant

Construct a new dynamic constant lookup.

storage

Construct a new dynamic storage lookup.

storage_root

Construct a new dynamic storage lookup to the root of some entry.

tx

Construct a new dynamic transaction payload to submit to a node.

Type Definitions

DecodedValue

A scale_value::Value type endowed with contextual information regarding what type was used to decode each part of it. This implements crate::metadata::DecodeWithMetadata, and is used as a return type for dynamic requests.