Type Definition subxt::dynamic::DecodedValue

pub type DecodedValue = Value<TypeId>;

A scale_value::Value type endowed with contextual information regarding what type was used to decode each part of it. This implements crate::metadata::DecodeWithMetadata, and is used as a return type for dynamic requests.