Module subxt::error

Types representing the errors that can be returned.

Structs

ModuleError

Details about a module error that has occurred.

ModuleErrorData

The error details about a module error that has occurred.

RpcError

An RPC error. Since we are generic over the RPC client that is used, the error is any custom string.

Enums

DecodeError

An error decoding SCALE bytes.

DispatchError

This is our attempt to decode a runtime DispatchError. We either successfully decode it into a ModuleError, or we fail and keep hold of the bytes, which we can attempt to decode if we have an appropriate static type to hand.

EncodeError

An error encoding a Value into SCALE bytes.

Error

The underlying error enum, generic over the type held by the Runtime variant. Prefer to use the Error<E> and Error aliases over using this type directly.

InvalidMetadataError

Error originated from converting a runtime metadata RuntimeMetadataPrefixed to the internal Metadata representation.

MetadataError

Metadata error originated from inspecting the internal representation of the runtime metadata.

SecretStringError

An error with the interpretation of a secret.

StorageAddressError

Something went wrong trying to encode a storage address.

TransactionError

Transaction error.

TransactionValidityError

Errors that can occur while checking the validity of a transaction.