Module subxt::metadata

Types representing the metadata obtained from a node.

Structs

CallData

Decoded call data and associated type information.

DecodeStaticType

Any type implementing Decode can also be decoded with the help of metadata.

Decoder

Decoder

DecoderBuilder

Pallet name

EncodeStaticType

Any type implementing Encode can also be encoded with the help of metadata.

ErrorMetadata

Details about a specific runtime error.

EventMetadata

Metadata for specific events.

Extrinsic

The result of successfully decoding an extrinsic.

Metadata

A representation of the runtime metadata received from a node.

MetadataPalletCalls

Metadata pallet calls

PalletMetadata

Metadata for a specific pallet.

PathKey

Path key

Enums

InvalidMetadataError

Error originated from converting a runtime metadata RuntimeMetadataPrefixed to the internal Metadata representation.

MetadataError

Metadata error originated from inspecting the internal representation of the runtime metadata.

Traits

DecodeWithMetadata

This trait is implemented for types which can be decoded with the help of metadata.

EncodeWithMetadata

This trait is implemented for types which can be encoded with the help of metadata.

MetadataLocation

Locate an item of a known type in the metadata. We should already know that the item we’re looking for is a call or event for instance, and then with this, we can dig up details for that item in the metadata.