Module subxt::rpc

RPC types and client for interacting with a substrate node.

These is used behind the scenes by various subxt APIs, but can also be used directly.

Rpc is the highest level wrapper, and the one you will run into first. It contains the higher level methods for interacting with a node.

RpcClient is what Rpc uses to actually talk to a node, offering a RpcClient::request and RpcClient::subscribe method to do so.

RpcClientT is the underlying dynamic RPC implementation. This provides the low level RpcClientT::request_raw and RpcClientT::subscribe_raw methods. This can be swapped out for a custom implementation, but by default we’ll rely on jsonrpsee for this.

Example

Fetching storage keys

use subxt::{ PolkadotConfig, OnlineClient, storage::StorageKey };

#[subxt::subxt(runtime_metadata_path = "../artifacts/polkadot_metadata.scale")]
pub mod polkadot {}

let api = OnlineClient::<PolkadotConfig>::new().await.unwrap();

let key = polkadot::storage()
    .xcm_pallet()
    .version_notifiers_root()
    .to_bytes();

// Fetch up to 10 keys.
let keys = api
    .rpc()
    .storage_keys_paged(&key, 10, None, None)
    .await
    .unwrap();

for key in keys.iter() {
    println!("Key: 0x{}", hex::encode(&key));
}

Macros

rpc_params

Create some RpcParams to pass to our RpcClient. RpcParams simply enforces that parameters handed to our RpcClient methods are the correct shape.

Structs

BlockNumber

Wrapper for NumberOrHex to allow custom From impls

BlockStats

Statistics of a block returned by the dev_getBlockStats RPC.

Health

Health struct returned by the RPC

RawValue

Reference to a range of bytes encompassing a single valid JSON value in the input data.

ReadProof

ReadProof struct returned by the RPC

Rpc

Client for substrate rpc interfaces

RpcClient

A concrete wrapper around an RpcClientT which exposes the udnerlying interface via some higher level methods that make it a little easier to work with.

RpcParams

This represents the parameters passed to an RpcClient, and exists to enforce that parameters are provided in the correct format.

RuntimeVersion

This contains the runtime version information necessary to make transactions, as obtained from the RPC call state_getRuntimeVersion,

Subscription

A generic RPC Subscription. This implements Stream, and so most of the functionality you’ll need to interact with it comes from the StreamExt extension trait.

TryFromIntError

An error type that signals an out-of-range conversion attempt.

Enums

NumberOrHex

A number type that can be serialized both as a number or a string that encodes a number in a string.

SubstrateTxStatus

Possible transaction status events.

Traits

RpcClientT

Any RPC client which implements this can be used in our super::Rpc type to talk to a node.

Type Definitions

ChainBlock

Alias for the type of a block returned by chain_getBlock

RpcFuture

A boxed future that is returned from the RpcClientT methods.

RpcSubscription

The inner subscription stream returned from our RpcClientT’s subscription method.

SystemProperties

Arbitrary properties defined in the chain spec as a JSON object.