Module subxt::tx

Create and submit extrinsics.

An extrinsic is submitted with an “signed extra” and “additional” parameters, which can be different for each chain. The trait crate::config::ExtrinsicParams determines exactly which additional and signed extra parameters are used when constructing an extrinsic, and is a part of the chain configuration (see crate::config::Config).

Structs

DynamicTxPayload

This represents a dynamically generated transaction payload.

PairSigner

A Signer implementation that can be constructed from an sp_core::Pair.

StaticTxPayload

This represents a statically generated transaction payload.

SubmittableExtrinsic

This represents an extrinsic that has been signed and is ready to submit.

TxClient

A client for working with transactions.

TxInBlock

This struct represents a transaction that has made it into a block.

TxProgress

This struct represents a subscription to the progress of some transaction.

Enums

TxStatus

Possible transaction statuses returned from our TxProgress::next_item() call.

Traits

Signer

Signing transactions requires a Signer. This is responsible for providing the “from” account that the transaction is being signed by, as well as actually signing a SCALE encoded payload.

TxPayload

This represents a transaction payload that can be submitted to a node.

Functions

dynamic

Construct a new dynamic transaction payload to submit to a node.